Move LogLevel to mod options
