+++ src/ops/mod.rs
@@ -15,10 +15,10 @@ use std::default::Default;
 use rand::{Rng, thread_rng};
 use iron::modifiers::Header;
 use iron::url::Url as GenericUrl;
-use self::super::{Options, Error};
 use mime_guess::guess_mime_type_opt;
 use hyper_native_tls::NativeTlsServer;
 use std::collections::{BTreeMap, HashMap};
+use self::super::{LogLevel, Options, Error};
 use std::io::{self, SeekFrom, Write, Read, Seek};
 use std::process::{ExitStatus, Command, Child, Stdio};
 use rfsapi::{RawFsApiHeader, FilesetData, RawFileData};
@@ -1275,30 +1275,6 @@ impl Clone for HttpHandler {
             proxies: self.proxies.clone(),
             cache_gen: Default::default(),
             cache_fs: Default::default(),
-        }
-    }
-}
-
-
-#[derive(Debug, Copy, Clone, Hash, PartialEq, Eq, PartialOrd, Ord)]
-pub enum LogLevel {
-    /// Write everything
-    All,
-    /// No serving messages
-    NoServeStatus,
-    /// No startup messages, but yes auth data
-    NoStartup,
-    /// No auth data
-    NoAuth,
-}
-
-impl From<u64> for LogLevel {
-    fn from(raw: u64) -> LogLevel {
-        match raw {
-            0 => LogLevel::All,
-            1 => LogLevel::NoServeStatus,
-            2 => LogLevel::NoStartup,
-            _ => LogLevel::NoAuth,
         }
     }
 }
